[Qt] Adjust size of splash screen image.
Add extra height to avoid having the initialization status text
overlap with the copyright text.

Since the splash window size is based on the static width/height of
the pre-rendered image files, just add an extra 10px to the image
height.

Also made the background for the testnet splash image uniform in
color.