[Qt] Fix segfault when running with `-help`
When running pivx-qt with the `-help` flag, the intended operation
is to output the help contents then exit cleanly. However, this is
currently resulting in a segfault due to the return status of a
method in intro.

This Fixes the segfaulting behavior and now cleanly exits after
displaying the help contents.