zPiv mints with < 3 mints added to accumulator after are immature.
To generate a witness, and therefore spend, a zPiv mint must have at least 3 mints added to the accumulator after it is accumulated itself. If that condition is not met, then the spender would receive the error message "Less than 3 mints added, unable to create spend".

Now, ListMintedCoins() will check to see if the mint has at least 3 mints after it before considering it mature and ready for spending.