Do not stringify anything beyond OP_ZEROCOINSPEND for CScript::ToString()
There are no further op codes in zerocoinspends. When parsing through the data, the ToString() method often see's the zPiv spend data as an OP code. Since RPC getrawtransaction prints the hex of the CScript anyways, it is redundant to print all of the spend data in this to string and then again in the hex dump.