Revert #402 "Change git info in genbuild.sh"
The previous change required tagging on `master`, which we don't do. As
a result, the version string displayed in-wallet was always displaying
based off of a dummy tag that was pushed during the review of #402.

This PR restores the functionality to generate the version string
without relying on superficial tags.

This reverts commit 8fa5f125e558bc4951e8af0ecea4f066813e4a66.