[Wallet] Add a check on zPIV spend to avoid a segfault
When zPIV coins are selected and their total value is less than the
target value of the payment, there was nothing to see it before
commiting the transaction. At which point the function would crap out
with an exception which QT would fail to handle, leading to a
segmentation fault. This new check avoids this situation.