[Qt] Fix a windows only crash when r-clicking a proposal
The problem seems to be related to a stack corruption issue
occuring within Qt and only for windows 64 builds. Further
investigation is required but this is a suitable fix for now
and there is no need to have the r-click doing the sae job as
the l-click in this situation.